moving nav-routines to new module
