+++ sw/airborne/modules/nav/nav_finken.h
@@ -32,4 +32,9 @@
 #include "std.h"
 #include "paparazzi.h"
 
+#include "modules/sonar/sonar_array_i2c.h"
+
+extern float sonar_failsave_pitch( void );
+extern float sonar_failsave_roll( void );
+
 #endif
