+++ sw/airborne/modules/sonar/sonar_array_i2c.c
@@ -84,8 +84,6 @@ uint8_t sonar_index;
 struct sonar_values_s sonar_values;
 struct sonar_values_s sonar_values_old;
 struct sonar_data_available_s sonar_data_available;
-struct sonar_values_s sonar_errors;
-struct sonar_values_s sonar_errors_old;
 #define SONAR_STATUS_IDLE 0
 #define SONAR_STATUS_PENDING 1
 
@@ -123,17 +121,6 @@ void sonar_array_i2c_init(void) {
 	sonar_values.left  = 0;
 	sonar_values.down  = 0;
 
-	sonar_errors.front = 0;
-	sonar_errors.right = 0;
-	sonar_errors.back  = 0;
-	sonar_errors.left  = 0;
-	sonar_errors.down  = 0;
-
-	sonar_errors_old.front = 0;
-	sonar_errors_old.right = 0;
-	sonar_errors_old.back  = 0;
-	sonar_errors_old.left  = 0;
-	sonar_errors_old.down  = 0;
 	// register telemetry
 	register_periodic_telemetry(DefaultPeriodic, "SONAR_ARRAY", send_sonar_array_telemetry);
 }
@@ -184,263 +171,6 @@ void sonar_array_i2c_periodic(void) {
 #else // SITL
 #warn "SITL not implemented for sonar_array_i2c yet"
 #endif // SITL
-}
-
-#ifndef SONAR_FAILSAVE_RANGE
-#define SONAR_FAILSAVE_RANGE 200
-#endif
-
-#ifndef SONAR_FAILSAVE_P
-#define SONAR_FAILSAVE_P 0.4
-#endif
-
-#ifndef SONAR_FAILSAVE_D
-#define SONAR_FAILSAVE_D 0.1
-#endif
-
-#ifndef SONAR_FAILSAVE_I
-#define SONAR_FAILSAVE_I 0.0
-#endif
-
-/////// Throttel
-#ifndef SONAR_FAILSAVE_P_DOWN
-#define SONAR_FAILSAVE_P_DOWN 0.4
-#endif
-
-#ifndef SONAR_FAILSAVE_D_DOWN
-#define SONAR_FAILSAVE_D_DOWN 0.1
-#endif
-
-#ifndef SONAR_FAILSAVE_I_DOWN
-#define SONAR_FAILSAVE_I_DOWN 0.0
-#endif
-
-double e_front_sum=0;
-double e_front_old=0;
-double e_back_sum=0;
-double e_back_old=0;
-
-double e_left_sum=0;
-double e_left_old=0;
-
-double e_right_sum=0;
-double e_right_old=0;
-
-double e_down_sum=0;
-double e_down_old=0;
-
-
-float sonar_failsave_pitch( void ) {
-	/*
-		 if (sonar_values.front > sonar_values.back) {
-		 float out = ( (SONAR_FAILSAVE_RANGE - sonar_values.back) * SONAR_FAILSAVE_P - (sonar_values_old.back - sonar_values.back) * SONAR_FAILSAVE_D);
-		 if(out > 0)
-		 return -out;
-		 return 0.0;
-		 } else {
-		 float out = ( (SONAR_FAILSAVE_RANGE - sonar_values.front) * SONAR_FAILSAVE_P - (sonar_values_old.front - sonar_values.front) * SONAR_FAILSAVE_D);
-		 if(out > 0)
-		 return out;
-		 return 0.0;
-		 }
-
-		 sonar_errors.front = SONAR_FAILSAVE_RANGE - sonar_values.front;
-		 if(sonar_errors.front < 0) {
-		 sonar_errors.front = 0;
-		 }
-
-		 sonar_errors.back = SONAR_FAILSAVE_RANGE - sonar_values.back;
-		 if(sonar_errors.back < 0) {
-		 sonar_errors.back = 0;
-		 }
-
-		 sonar_errors_old.front = SONAR_FAILSAVE_RANGE - sonar_values.front;
-		 if(sonar_errors_old.front < 0) {
-		 sonar_errors_old.front = 0;
-		 }
-
-		 sonar_errors_old.back = SONAR_FAILSAVE_RANGE - sonar_values.back;
-		 if(sonar_errors_old.back < 0) {
-		 sonar_errors_old.back = 0;
-		 }
-
-		 float e = sonar_errors.front - sonar_errors.back;
-		 float de = e - sonar_errors_old.front - sonar_errors_old.back;
-
-		 float y = e * SONAR_FAILSAVE_P - de * SONAR_FAILSAVE_D;
-		 return y;
-		 */
-
-	// New PID control
-
-	// Sonar Winkel korrektur
-	//doule distanz_correct=0;
-	//distanz_correct=cos(pitchangle)*sonar_values.front
-
-	double saveDistanz=100;
-	double Ta=0.166;	
-
-	// Distanz front
-	double e_front= (saveDistanz - sonar_values.front);
-	e_front_sum = e_front_sum + e_front;
-	e_front_old = (saveDistanz - sonar_values_old.front);
-
-	float out_front = e_front * SONAR_FAILSAVE_P + 
-		((e_front - e_front_old) * SONAR_FAILSAVE_D / Ta) + SONAR_FAILSAVE_I * Ta * e_front_sum;
-
-	if(sonar_values.front>120)
-	{
-		out_front=0;
-		e_front_sum=0;
-	}
-
-	// Distanz back
-	double e_back= (saveDistanz - sonar_values.back);
-	e_back_sum = e_back_sum + e_back;
-	e_back_old = (saveDistanz - sonar_values_old.back);
-
-
-	float out_back = e_back * SONAR_FAILSAVE_P + 
-		((e_back - e_back_old) * SONAR_FAILSAVE_D / Ta) + SONAR_FAILSAVE_I * Ta * e_back_sum;
-
-	if(sonar_values.back>120)
-	{
-		out_back=0;
-		e_back_sum=0;
-	}
-
-	// Differenz between back and front
-	double pitchangle=0;
-	if(sonar_values.front<=120 || sonar_values.back<=120)
-	{
-		pitchangle=out_front-out_back;
-	}
-
-	return pitchangle;
-}
-
-float sonar_throttel( void )
-{
-	double saveDistanz=100;
-	double Ta=0.066;	
-
-	// Distanz down
-	double e_down= (saveDistanz - sonar_values.down);
-	e_down_sum = e_down_sum + e_down;
-
-	float out_down = e_down * SONAR_FAILSAVE_P_DOWN + 
-		((e_down - e_down_old) * SONAR_FAILSAVE_D_DOWN / Ta) + SONAR_FAILSAVE_I_DOWN * Ta * e_down_sum;
-
-	e_down_old= e_down;
-
-	if(sonar_values.down<21)
-	{
-		out_down=0;
-		e_down_sum=0;
-		e_down_old=e_down;
-	}
-
-
-	float throttel=0;
-	throttel=out_down+0.42;
-
-	if(throttel>0.7)
-	{
-		return 0.7;
-	}
-	if(throttel<0.0)
-	{
-		return 0.0;
-	}
-
-
-	return throttel;
-
-}
-
-
-float sonar_failsave_roll( void ) {
-	/*
-		 if (sonar_values.left > sonar_values.right) {
-		 float out = ( (SONAR_FAILSAVE_RANGE - 1.0 * sonar_values.right) * SONAR_FAILSAVE_P  - (sonar_values_old.right - sonar_values.right) * SONAR_FAILSAVE_D); 
-		 if(out > 0)
-		 return -out;
-		 return 0.0;
-		 } else {
-		 float out = ( (SONAR_FAILSAVE_RANGE - 1.0 * sonar_values.left) * SONAR_FAILSAVE_P - (sonar_values_old.left - sonar_values.left) * SONAR_FAILSAVE_D); 
-		 if(out > 0)
-		 return out;
-		 return 0.0;
-
-		 sonar_errors.right = SONAR_FAILSAVE_RANGE - sonar_values.right;
-		 if(sonar_errors.right < 0) {
-		 sonar_errors.right = 0;
-		 }
-
-		 sonar_errors.left = SONAR_FAILSAVE_RANGE - sonar_values.left;
-		 if(sonar_errors.left < 0) {
-		 sonar_errors.left = 0;
-		 }
-
-		 sonar_errors_old.right = SONAR_FAILSAVE_RANGE - sonar_values.right;
-		 if(sonar_errors_old.right < 0) {
-		 sonar_errors_old.right = 0;
-		 }
-
-		 sonar_errors_old.left = SONAR_FAILSAVE_RANGE - sonar_values.left;
-		 if(sonar_errors_old.left < 0) {
-		 sonar_errors_old.left = 0;
-		 }
-
-		 float e = sonar_errors.right - sonar_errors.left;
-		 float de = e - sonar_errors_old.right - sonar_errors_old.left;
-
-		 float y = -1.0 * (e * SONAR_FAILSAVE_P - de * SONAR_FAILSAVE_D);
-		 y=0;
-		 return y;
-		 */
-	double saveDistanz=100;
-	double Ta=0.05;	
-
-	// Distanz left
-	double e_left= (saveDistanz - sonar_values.left);
-	e_left_sum = e_left_sum + e_left;
-
-	float out_left = e_left * SONAR_FAILSAVE_P + 
-		((e_left - e_left_old) * SONAR_FAILSAVE_D / Ta) + SONAR_FAILSAVE_I * Ta * e_left_sum;
-
-	e_left_old= e_left;
-
-	if(sonar_values.left>120)
-	{
-		out_left=0;
-		e_left_sum=0;
-		e_left_old=e_left;
-	}
-
-	// Distanz right
-	double e_right= (saveDistanz - sonar_values.right);
-	e_right_sum = e_right_sum + e_right;
-
-	float out_right = e_right * SONAR_FAILSAVE_P + 
-		((e_right - e_right_old) * SONAR_FAILSAVE_D / Ta) + SONAR_FAILSAVE_I * Ta * e_right_sum;
-
-	e_right_old= e_right;
-	if(sonar_values.right>120)
-	{
-		out_right=0;
-		e_right_sum=0;
-		e_right_old=e_right;
-	}
-
-	// Differenz between back and front
-	double roll=0;
-	if(sonar_values.left<=120 || sonar_values.right<=120)
-	{
-		roll=out_left-out_right;
-	}
-
-	return roll;
 }
 void sonar_array_i2c_event( void ) { 
 	// i guess it it not possible to query verything so often
